requirements-dev.txt
# Test dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-httpx>=0.25.0

# Code quality
flake8>=4.0.0
black>=22.0.0
isort>=5.0.0
mypy>=0.990

# Runtime dependencies (for testing)
httpx>=0.25.0
pydantic>=1.10.0

# For running examples
python-dotenv>=0.19.0
